rendered everything on a single page
we need to set routing
