+++ src/App.js
@@ -1,5 +1,7 @@
 import React, { useState } from "react";
 import "./App.css";
+import Profile from './profile.jsx';
+import DataPage from './DataPage.jsx';
 
 import Login from './components/Authentication/login';
 import Signup from './components/Authentication/signup';
@@ -8,10 +10,13 @@ import AudioSender from "./components/AudioRecorder/AudioSender";
 function App() {
   return (
     <div className="App">
-      {/* <Signup /> */}  
+      {/* <Signup /> */}
       {/* <Login /> */}
       {/* <AudioRecorder /> */}
       <AudioSender/>
+      <Profile/>
+      <DataPage />
+
       {/* <SendAudio /> */}
     </div>
   );
